added and tested  auth

+++ courses/views.py
@@ -4,9 +4,12 @@ from rest_framework.response import Response
 from .models import Course
 from .serializers.common import CourseSerializer
 from utils.decorators import handle_exceptions
+from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly, IsAdminUser
+from rest_framework.exceptions import PermissionDenied
 # Create your views here.
 
 class CourseListCreateView(APIView):
+    permission_classes = [IsAuthenticated or IsAdminUser]
     @handle_exceptions
     def get(self, request):
         courses= Course.objects.all()
@@ -16,15 +19,16 @@ class CourseListCreateView(APIView):
 
     @handle_exceptions
     def post(self, request):
+            request.data['owner']= request.user.id
             course_to_create = CourseSerializer(data = request.data)
             if course_to_create.is_valid():
                 course_to_create.save()
                 return Response(course_to_create.data, 201)
-            print("validation error:", course_to_create.error)
+            print("validation error:", course_to_create.errors)
             return Response(course_to_create.errors, 400)
 
 class CourseRetrieveUpdateDestroyView(APIView):
-
+    permission_classes = [IsAuthenticated or IsAdminUser]
     @handle_exceptions
     def get(self, request, pk):
         course = Course.objects.get(pk=pk)
@@ -35,6 +39,11 @@ class CourseRetrieveUpdateDestroyView(APIView):
     @handle_exceptions
     def put(self, request, pk):
         course_to_update = Course.objects.get(pk=pk)
+        #! here
+        if course_to_update.owner != request.user and request.user.is_staff == False:
+            #print('on fail req user', request.user.is_staff, course_to_update.owner)
+            raise PermissionDenied()
+        #print('on succ req user', request.user.is_staff, course_to_update.owner)
         serialized_course= CourseSerializer(course_to_update, data = request.data, partial=True)
         if serialized_course.is_valid():
             serialized_course.save()
@@ -44,7 +53,10 @@ class CourseRetrieveUpdateDestroyView(APIView):
     @handle_exceptions
     def delete(self, request, pk):
         course_to_delete= Course.objects.get(pk=pk)
-        print('Deleteing:', course_to_delete)
+        if course_to_delete.owner != request.user and request.user.is_staff == False:
+            #print('on fail req user', request.user.is_staff, course_to_delete.owner)
+            raise PermissionDenied()
+        print('Deleting:', course_to_delete)
         course_to_delete.delete()
         return Response(status=204)
